# Performance improvement in contains_algorithm