@quantumsteve 2020-04-21
These changes may make a difference if installed on a network filesystem.

spaced timings out by 15 minutes to reduce the likelihood files are in cache. (Unfortunately I don't have sufficient privileges on this workstation to do better)

before: 20.935s, 11.431s, 14.518s, 14.767s, 15.298s, 15.571s
after: 7.950s, 7.865s,  7.530s, 7.941s, 8.751s, 9.650s